Steganografija/requirements.txt
Pillow>=9.0.0
cryptography>=41.0.0
